move functions in `hydro_dem_utils.py` to `process_hydro_dem.py`
